Download All Station Data

This function downloads all of the hourly station data across all avalible stations for every year that data is avalible.

download_all_station_data(
  testing = FALSE,
  limit_for_testing_purposes,
  temporary = FALSE
)

Arguments

testing

this determines if the url will actually be downloaded, or if we are just pretending for the sake of both my sanity and my computer. Defults to FALSE.

limit_for_testing_purposes

because I want to show that it actually works, you can limit the number of stations that you want, and it will pick that many random stations to download in their entirety.

temporary

this determines if the files are temporary or if they are saved to the computer. Defaults to FALSE.

Value

if temporary is set to TRUE, this (theoretically) returns a giant data frame of all of the hourly station data from all available years and all available stations. Or apparently as a list of data.frames, which I'm not sure why it's doing that and am too tired to figure out. If temporary is set to FALSE, those files will be downloaded in subdirectories named after their associated station.

Examples

full_station_data <- download_all_station_data(limit_for_testing_purposes = 2, temporary = TRUE)
#> You are about to download a metric butload of files. Are you sure you want to do this? (Yes/no/cancel) 
#> Are you 100% sure? This may take a few days just so you know. I haven't actually used it to download all the files so I'm not sure the exact amount of time (Yes/no/cancel) 
#> Error in {    Datacleaning::download_all_one_station(station_ID = station_id,         testing = testing, temporary = temporary)}: task 1 failed - "cannot open the connection"
str(full_station_data)
#> Error: object 'full_station_data' not found